Bugfix: Extra bot config file copy
- Exception: NoSuchFileException
- Reproducing the bug:
  - Load a ZIP file containing a bot and an extra config file.
  - Click "Start" to connect the bot.
  - Click "Eject" to stop the bot.
  - Load a ZIP file containing a different bot with no config files.
  - Click "Start" and a "NoSuchFileException" would be thrown
  for the previous bot's config file.
- The above bug has been fixed with this commit.

+++ src/droplauncher/bwheadless/BWHeadless.java
@@ -475,7 +475,9 @@ public class BWHeadless {
 
     /* Copy extra files to common bot I/O directories. */
     for (String path : this.bot.getExtraFiles()) {
-      Files.copy(Paths.get(path), Paths.get(bwapiAiPath.toString(), FilenameUtils.getName(path)), StandardCopyOption.REPLACE_EXISTING);
+      if (AdakiteUtils.fileExists(Paths.get(path))) {
+        Files.copy(Paths.get(path), Paths.get(bwapiAiPath.toString(), FilenameUtils.getName(path)), StandardCopyOption.REPLACE_EXISTING);
+      }
     }
   }
